Test caches for linked resources
Make sure linked resources are being loaded from resources caches if
they exist.

Some links are only defined by a URL so this double caches resources by
id and by their href element if it exists. This takes care of caching
avatars if they're queried for using ?include=avatar

Previously the code to fetch a linked resource was busting the resource
cache by passing an additional non-ID argument 'query' that was usually
undefined. This forced the client to re-request for the resource instead
of loading it from the cache by id.